TOTS-62 More refactoring
